+++ CHANGELOG.md
@@ -18,6 +18,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3642 Hide the 'basal safety profile' menu for mobigo+ users
 - YLP-3651 Transfer the patient information displayed in the banner in the patient profile page
 - YLP-3687 Update insulin stats order
+- YLP-3690 Update basal tooltip style in Daily view
 
 ### Engineering
 - YLP-3672 D3 version update